deep-jass-project/requirements.txt
certifi==2021.5.30
setuptools==58.0.4
flask==2.0.1
jass-kit==2.0.3
numpy==1.21.2
gunicorn==20.1.0
bitcoinaddress==0.1.5
matplotlib==3.5.0
pandas==1.3.4
sklearn==0.0
tensorflow==2.7.0
futures==2.2.0